make pi2/pizero LED support more consistent with previous similar support
